mandatory code cleanup
