--- src/components/Base/Link/Link.classnames.ts
@@ -1,19 +1,0 @@
-import { type VariantProps, cva } from "class-variance-authority";
-
-export const LinkCVA = cva("link underline-offset-2", {
-  variants: {
-    colorStyle: {
-      primary: "text-primary",
-      secondary: "text-secondary",
-    },
-    isBold: {
-      true: "font-bold",
-    },
-    defaultVariants: {
-      isBold: false,
-      colorStyle: "primary",
-    },
-  },
-});
-
-export type LinkVariantProps = NonNullable<VariantProps<typeof LinkCVA>>;
