--- src/components/Base/index.ts
@@ -1,4 +1,0 @@
-export { default as Button } from "./Button/Button";
-export { default as Text } from "./Text/Text";
-export { default as Link } from "./Link/Link";
-export { default as Badge } from "./Badge/Badge";
