+++ src/components/index.ts
@@ -1,0 +1,3 @@
+export { default as Button } from "./Button";
+export { default as Text } from "./Text";
+export { default as Badge } from "./Badge";
