+++ src/layouts/About.astro
@@ -1,9 +1,9 @@
 ---
-import { Badge, Text } from "./Base";
 import Container from "./Container.astro";
 import DataDrivenIcon from "../../public/icons/DataDriven.svg";
 import TheCurveIcon from "../../public/icons/Curve.svg";
 import VisibilityIcon from "../../public/icons/Visibility.svg";
+import { Text } from "../components";
 
 const sectionTitle = "What we are all about";
 const sectionDescription =
@@ -64,3 +64,4 @@ const values = [
     </div>
   </div>
 </Container>
+ ../components
