+++ src/layouts/CTA.astro
@@ -1,5 +1,5 @@
 ---
-import { Badge, Button, Text } from "./Base";
+import { Button, Text } from "../components";
 import Container from "./Container.astro";
 const sectionTitle = "Connect with Our Team for Tailored Digital Solutions";
 const sectionDescription =
