+++ src/layouts/Footer.astro
@@ -1,5 +1,5 @@
 ---
-import { Text } from "./Base";
+import { Text } from "../components";
 ---
 
 <footer class="border-t border-base-300 mt-xl lg:mt-3xl">
