+++ src/layouts/Hero.astro
@@ -1,7 +1,6 @@
 ---
-import { Text, Button, Badge } from "./Base/index";
+import { Text, Button, Badge } from "../components";
 import Container from "./Container.astro";
-import { BoltIcon } from "@heroicons/react/24/outline";
 ---
 
 <Container>
