+++ src/layouts/LogoCloud.astro
@@ -1,5 +1,5 @@
 ---
-import { Text } from "./Base";
+import { Text } from "../components";
 import Container from "./Container.astro";
 ---
