+++ src/layouts/Nav.astro
@@ -1,5 +1,5 @@
 ---
-import { Button } from "./Base";
+import { Button } from "../components";
 ---
 
 <nav class="navbar">
