+++ src/layouts/Process.astro
@@ -1,5 +1,5 @@
 ---
-import { Badge, Text } from "./Base";
+import { Badge, Text } from "../components";
 import Container from "./Container.astro";
 
 const sectionTitle = "A simple, yet powerful and efficient process";
