+++ src/layouts/Services.astro
@@ -1,5 +1,5 @@
 ---
-import { Badge, Text } from "./Base";
+import { Badge, Text } from "../components";
 import PPC from "../../public/icons/PPC.svg";
 import CRO from "../../public/icons/CRO.svg";
 import Content from "../../public/icons/Content.svg";
