+++ src/layouts/Testimonials.astro
@@ -1,0 +1,19 @@
+---
+import { Text } from "../components";
+---
+
+<section class="py-md lg:py-lg">
+  <div class="max-w-full flex justify-center">
+    <Text variant="h1" className="w-full lg:w-2/3" align="center">
+      Check our clients feedback!
+    </Text>
+  </div>
+  <div
+    class="grid grid-cols-1 gap-lg pt-xl md:grid-cols-2 lg:grid-cols-3 lg:gap-xl lg:pt-2xl"
+  >
+    <div class="p-3xl rounded-lg bg-neutral-200"></div>
+    <div class="p-3xl rounded-lg bg-neutral-200"></div>
+    <div class="p-3xl rounded-lg bg-neutral-200"></div>
+  </div>
+</section>
+ ../components
