+++ src/pages/index.astro
@@ -1,10 +1,10 @@
 ---
-import About from "../components/About.astro";
-import CTA from "../components/CTA.astro";
-import Hero from "../components/Hero.astro";
-import MainLayout from "../components/MainLayout.astro";
-import Process from "../components/Process.astro";
-import Services from "../components/Services.astro";
+import About from "../layouts/About.astro";
+import CTA from "../layouts/CTA.astro";
+import Hero from "../layouts/Hero.astro";
+import MainLayout from "../layouts/MainLayout.astro";
+import Process from "../layouts/Process.astro";
+import Services from "../layouts/Services.astro";
 ---
 
 <MainLayout title="Toggle Solutions | Home">
